% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\alias{explain.nn_module}
\alias{explain.extracted_survivalmodels_coxtime}
\alias{explain.extracted_survivalmodels_deephit}
\alias{explain.extracted_survivalmodels_deepsurv}
\title{Explain a model}
\usage{
explain(
  model,
  data = NULL,
  model_type = NULL,
  baseline_hazard = NULL,
  labtrans = NULL,
  time_bins = NULL,
  preprocess_fun = NULL,
  postprocess_fun = NULL,
  predict_fun = NULL
)

\method{explain}{nn_module}(
  model,
  data,
  model_type,
  baseline_hazard = NULL,
  labtrans = NULL,
  time_bins = NULL,
  preprocess_fun = NULL,
  postprocess_fun = NULL,
  predict_fun = NULL
)

\method{explain}{extracted_survivalmodels_coxtime}(model, data, ...)

\method{explain}{extracted_survivalmodels_deephit}(model, data, ...)

\method{explain}{extracted_survivalmodels_deepsurv}(model, data, ...)
}
\arguments{
\item{model}{A model object.}

\item{data}{A data frame or matrix of data to explain the model.}

\item{model_type}{A string specifying the type of the survival model.
Possible values are "coxtime", "deephit", or "deepsurv".}

\item{baseline_hazard}{A data frame containing the baseline hazard. It
should have two columns: "time" and "hazard". This is only used for
"coxtime" and "deepsurv" models.}

\item{labtrans}{A list containing the transformation functions for the
time variable. It should have two elements: "transform" and
"transform_inv". This is highly experimental and not yet fully
supported.}

\item{time_bins}{A numeric vector specifying the time bins for the
"deephit" model, e.g., \code{c(0, 1, 2, 3)}.}

\item{preprocess_fun}{A function to preprocess the data before
making predictions, e.g., adding a time variable for a \code{coxtime} model.
This argument is highly experimental and the default values should work
for most cases.}

\item{postprocess_fun}{A function to postprocess the predictions after
making them. This argument is highly experimental and the default
values should work.}

\item{predict_fun}{A function that can be used to make predictions
from the model. If \code{NULL}, the predict method of the model will be used.}

\item{...}{Unused arguments.}
}
\value{
An object of class \code{explainer} that contains the model, the
data, and the prediction function.
}
\description{
This function is a generic method that dispatches to the appropriate
explain method based on the class of the model.
}
