% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_survivalmodels.R
\name{extract_model}
\alias{extract_model}
\alias{extract_model.coxtime}
\alias{extract_model.deephit}
\alias{extract_model.deepsurv}
\title{Extract model information from a \code{survivalmodels} object}
\usage{
extract_model(x, path = NULL)

\method{extract_model}{coxtime}(x, path = NULL)

\method{extract_model}{deephit}(x, path = NULL)

\method{extract_model}{deepsurv}(x, path = NULL)
}
\description{
Extract model information from a \code{survivalmodels} object
}
