% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtime.R
\name{load_pycox_model}
\alias{load_pycox_model}
\title{Load a pre-trained PyCox model}
\usage{
load_pycox_model(
  path,
  in_features,
  base_hazard,
  type = "CoxTime",
  architecture = NULL,
  num_nodes = c(32, 32),
  batch_norm = TRUE,
  dropout = 0.1,
  activation = "relu",
  output_activation = "linear"
)
}
\arguments{
\item{path}{The path to the weights of the PyCox model.}

\item{type}{The type of the model. Currently only "CoxTime" is supported.}

\item{architecture}{The architecture of the model. If NULL, the default
architecture is used. It needs to fit to the weights given in path.}
}
\description{
This function loads a pre-trained PyCox model from a given path.
}
