% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.explainer_coxtime}
\alias{print.explainer_coxtime}
\alias{print.explainer_deepsurv}
\alias{print.explainer_deephit}
\title{Custom print method for explainer objects}
\usage{
\method{print}{explainer_coxtime}(x, ...)

\method{print}{explainer_deepsurv}(x, ...)

\method{print}{explainer_deephit}(x, ...)
}
\arguments{
\item{x}{An object of class 'explainer_coxtime', 'explainer_deepsurv', or
'explainer_deephit'.}

\item{...}{Additional arguments (not used).}
}
\description{
This function prints a summary of the explainer object.
}
