% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.extracted_survivalmodels_coxtime}
\alias{print.extracted_survivalmodels_coxtime}
\alias{print.extracted_survivalmodels_deepsurv}
\alias{print.extracted_survivalmodels_deephit}
\title{Print method for extracted pycox survival model}
\usage{
\method{print}{extracted_survivalmodels_coxtime}(x, ...)

\method{print}{extracted_survivalmodels_deepsurv}(x, ...)

\method{print}{extracted_survivalmodels_deephit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{extracted_survivalmodels_coxtime},
\code{extracted_survivalmodels_deepsurv}, or \code{extracted_survivalmodels_deephit}.}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for extracted pycox survival model
}
