% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.surv_result}
\alias{print.surv_result}
\title{Print function for surv_result objects}
\usage{
\method{print}{surv_result}(x, ...)
}
\arguments{
\item{x}{An object of class "surv_result"}

\item{...}{Additional arguments (not used)}
}
\description{
Print function for surv_result objects
}
