% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_hessian.R
\name{surv_hessian}
\alias{surv_hessian}
\alias{surv_hessian.explainer_coxtime}
\title{Calculate the Gradient of the Survival Function}
\usage{
surv_hessian(
  exp,
  target = "survival",
  instance = 1,
  times_input = FALSE,
  batch_size = 50,
  dtype = "float",
  include_time = FALSE
)

\method{surv_hessian}{explainer_coxtime}(
  exp,
  target = "survival",
  instance = 1,
  times_input = FALSE,
  batch_size = 50,
  dtype = "float",
  include_time = FALSE
)
}
\description{
Calculate the Gradient of the Survival Function
}
\seealso{
Other Attribution Methods: 
\code{\link{surv_gradSHAP}()},
\code{\link{surv_grad}()},
\code{\link{surv_intgrad}()},
\code{\link{surv_inthess}()},
\code{\link{surv_smoothgrad}()}
}
\concept{Attribution Methods}
