% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_inthess.R
\name{surv_inthess}
\alias{surv_inthess}
\alias{surv_inthess.explainer_deepsurv}
\alias{surv_inthess.explainer_coxtime}
\alias{surv_inthess.explainer_deephit}
\title{Calculate the Integrated Hessian of the Survival Function}
\usage{
surv_inthess(
  exp,
  target = "survival",
  instance = 1,
  times_input = TRUE,
  batch_size = 50,
  n = 10,
  x_ref = NULL,
  dtype = "float",
  include_time = FALSE
)

\method{surv_inthess}{explainer_deepsurv}(
  exp,
  target = "survival",
  instance = 1,
  times_input = TRUE,
  batch_size = 50,
  n = 10,
  x_ref = NULL,
  dtype = "float",
  ...
)

\method{surv_inthess}{explainer_coxtime}(
  exp,
  target = "survival",
  instance = 1,
  times_input = TRUE,
  batch_size = 50,
  n = 10,
  x_ref = NULL,
  dtype = "float",
  include_time = FALSE
)

\method{surv_inthess}{explainer_deephit}(
  exp,
  target = "survival",
  instance = 1,
  times_input = TRUE,
  batch_size = 50,
  n = 10,
  x_ref = NULL,
  dtype = "float",
  ...
)
}
\description{
Calculate the Integrated Hessian of the Survival Function
}
\seealso{
Other Attribution Methods: 
\code{\link{surv_gradSHAP}()},
\code{\link{surv_grad}()},
\code{\link{surv_hessian}()},
\code{\link{surv_intgrad}()},
\code{\link{surv_smoothgrad}()}
}
\concept{Attribution Methods}
