% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survinng-package.R
\docType{package}
\name{survinng-package}
\alias{survinng}
\alias{survinng-package}
\title{survinng: Gradient-Based Feature Attribution for Survival Neural Networks}
\description{
This package implements model-specific, gradient-based feature attribution methods for deep survival neural networks, including DeepHit, CoxTime, and DeepSurv. It accompanies the ICML 2025 paper "Gradient-based Explanations for Deep Learning Survival Models".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bips-hb.github.io/survinng/}
}

}
\author{
\strong{Maintainer}: Niklas Koenen \email{niklas.koenen@gmail.com} (\href{https://orcid.org/0000-0002-4623-8271}{ORCID})

Authors:
\itemize{
  \item Sophie Hanna Langbein (\href{https://orcid.org/0000-0001-5629-2055}{ORCID})
}

}
\keyword{internal}
