% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{as.data.frame.surv_result}
\alias{as.data.frame.surv_result}
\alias{as.data.table.surv_result}
\title{Convert survival attribution results to a data frame}
\usage{
\method{as.data.frame}{surv_result}(x, ...)

as.data.table.surv_result(x, ...)
}
\description{
Convert survival attribution results to a data frame
}
\concept{Conversion Methods}
