% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{explain}
\alias{explain}
\alias{explain.nn_module}
\alias{explain.extracted_survivalmodels_coxtime}
\alias{explain.extracted_survivalmodels_deephit}
\alias{explain.extracted_survivalmodels_deepsurv}
\title{Explain a model}
\usage{
explain(
  model,
  data = NULL,
  model_type = NULL,
  baseline_hazard = NULL,
  labtrans = NULL,
  preprocess_fun = NULL,
  postprocess_fun = NULL,
  predict_fun = NULL
)

\method{explain}{nn_module}(
  model,
  data,
  model_type,
  baseline_hazard = NULL,
  labtrans = NULL,
  preprocess_fun = NULL,
  postprocess_fun = NULL,
  predict_fun = NULL
)

\method{explain}{extracted_survivalmodels_coxtime}(model, data, ...)

\method{explain}{extracted_survivalmodels_deephit}(model, data, ...)

\method{explain}{extracted_survivalmodels_deepsurv}(model, data, ...)
}
\arguments{
\item{model}{A model object.}

\item{data}{A data frame or matrix of data to explain the model.}

\item{predict_function}{A function that can be used to make predictions
from the model. If NULL, the predict method of the model will be used.}
}
\value{
An object of class \code{explainer} that contains the model, the
data, and the prediction function.
}
\description{
This function is a generic method that dispatches to the appropriate
explain method based on the class of the model.
}
