% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_survivalmodels.R
\name{extract_model}
\alias{extract_model}
\alias{extract_model.coxtime}
\alias{extract_model.deephit}
\alias{extract_model.deepsurv}
\title{Extract model information from a \code{survivalmodels} object}
\usage{
extract_model(x, path = NULL, num_basehazard = 200L)

\method{extract_model}{coxtime}(x, path = NULL, num_basehazard = 200L)

\method{extract_model}{deephit}(x, path = NULL, ...)

\method{extract_model}{deepsurv}(x, path = NULL, num_basehazard = 200L)
}
\arguments{
\item{x}{A \code{survivalmodels} object, i.e., \code{survivalmodels::deephit},
\code{survivalmodels::coxtime}, or \code{survivalmodels::deepsurv}.}

\item{path}{A string specifying the path to save the extracted model. If \code{NULL},
the model is not saved. Default is \code{NULL}.}

\item{num_basehazard}{An integer specifying the number of points of the
baseline hazard to compute. Default is \code{200L}. This argument is only used for
\code{coxtime} and \code{deepsurv} models.}

\item{...}{Unused arguments.}
}
\description{
This function extracts model information from a neural network trained with
the \code{survivalmodels} package.
}
