% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.surv_result}
\alias{plot.surv_result}
\title{Plot method for survival attribution results}
\usage{
\method{plot}{surv_result}(x, normalize = FALSE, add_sum = FALSE, ...)
}
\description{
Plot method for survival attribution results
}
\concept{Plot Methods}
