% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_grad.R
\name{surv_grad}
\alias{surv_grad}
\alias{surv_grad.explainer_deepsurv}
\alias{surv_grad.explainer_coxtime}
\alias{surv_grad.explainer_deephit}
\title{Calculate the Gradient of the Survival Function}
\usage{
surv_grad(
  exp,
  target = "survival",
  instance = 1,
  times_input = FALSE,
  batch_size = 50,
  include_time = FALSE
)

\method{surv_grad}{explainer_deepsurv}(
  exp,
  target = "survival",
  instance = 1,
  times_input = FALSE,
  batch_size = 50,
  ...
)

\method{surv_grad}{explainer_coxtime}(
  exp,
  target = "survival",
  instance = 1,
  times_input = FALSE,
  batch_size = 50,
  include_time = FALSE
)

\method{surv_grad}{explainer_deephit}(
  exp,
  target = "survival",
  instance = 1,
  times_input = FALSE,
  batch_size = 50,
  ...
)
}
\description{
Calculate the Gradient of the Survival Function
}
\seealso{
Other Attribution Methods: 
\code{\link{surv_gradSHAP}()},
\code{\link{surv_intgrad}()},
\code{\link{surv_smoothgrad}()}
}
\concept{Attribution Methods}
