% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_gradSHAP.R
\name{surv_gradSHAP}
\alias{surv_gradSHAP}
\alias{surv_gradSHAP.explainer_coxtime}
\alias{surv_gradSHAP.explainer_deephit}
\title{Calculate the GradSHAP values of the Survival Function}
\usage{
surv_gradSHAP(
  exp,
  target = "survival",
  instance = 1,
  n = 50,
  num_samples = 10,
  data_ref = NULL,
  use_base_hazard = TRUE,
  batch_size = 50
)

\method{surv_gradSHAP}{explainer_coxtime}(
  exp,
  target = "survival",
  instance = 1,
  n = 50,
  num_samples = 10,
  data_ref = NULL,
  use_base_hazard = TRUE,
  batch_size = 50
)

\method{surv_gradSHAP}{explainer_deephit}(
  exp,
  target = "survival",
  instance = 1,
  n = 50,
  num_samples = 10,
  data_ref = NULL,
  batch_size = 50
)
}
\description{
Calculate the GradSHAP values of the Survival Function
}
\seealso{
Other Attribution Methods: 
\code{\link{surv_grad}()},
\code{\link{surv_intgrad}()},
\code{\link{surv_smoothgrad}()}
}
\concept{Attribution Methods}
