% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_smoothgrad.R
\name{surv_smoothgrad}
\alias{surv_smoothgrad}
\alias{surv_smoothgrad.explainer_coxtime}
\alias{surv_smoothgrad.explainer_deephit}
\title{Calculate the SmoothGrad values of the Survival Function}
\usage{
surv_smoothgrad(
  exp,
  target = "survival",
  instance = 1,
  times_input = FALSE,
  n = 10,
  noise_level = 0.1,
  use_base_hazard = TRUE,
  include_time = FALSE,
  batch_size = 50
)

\method{surv_smoothgrad}{explainer_coxtime}(
  exp,
  target = "survival",
  instance = 1,
  times_input = FALSE,
  n = 10,
  noise_level = 0.1,
  use_base_hazard = TRUE,
  include_time = FALSE,
  batch_size = 50
)

\method{surv_smoothgrad}{explainer_deephit}(
  exp,
  target = "survival",
  instance = 1,
  times_input = FALSE,
  n = 10,
  noise_level = 0.1,
  use_base_hazard = TRUE,
  include_time = FALSE,
  batch_size = 50
)
}
\description{
Calculate the SmoothGrad values of the Survival Function
}
\seealso{
Other Attribution Methods: 
\code{\link{surv_gradSHAP}()},
\code{\link{surv_grad}()},
\code{\link{surv_intgrad}()}
}
\concept{Attribution Methods}
